requirements.txt
pre-commit==3.6.0
google-generativeai==0.3.1
gitpython==3.1.41
pyyaml==6.0.1
colorama==0.4.6 